requirements.txt
python-nmap>=0.7.1
dnspython>=2.4.0
requests>=2.31.0
